requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
matplotlib>=3.7.0
yfinance>=0.2.50
